[training, perf] refactor: Address review on eval-CP feature
Two review nits from yaoyu-33 on PR #3755:

1. Make the `train_pgs` / `eval_pgs` attributes on `GlobalState` private
   (underscore prefix), matching the convention already used by sibling
   internal-only attributes (`_nvrx_straggler_created`, `_energy_monitor`,
   etc.). The fields are only read by the eval-CP rebinding helper, not
   by general training code.

2. Rename `eval_cp.py` -> `eval_context_parallel_rebinding.py` so the
   module name is self-documenting. `eval_cp` was ambiguous (could be
   parsed as "evaluate cp" or "evaluation copy"); the new name reflects
   that the module's job is rebinding cached CP-group references at
   eval time.

Imports and the docstring reference in `training/config.py` updated to
the new module path. Public function name `eval_cp_context` is unchanged
so call sites in the example only needed the import line bumped.

No behavior change.
